finalizado o crud de livro (na teoria)

+++ src/main/java/repository/EditoraRepository.java
@@ -8,5 +8,5 @@ public interface EditoraRepository {
 
     void salvar(Editora editora);
     List<Editora> listarEditora();
-    Editora buscarEditora(Long id);
+    Editora buscarEditora(int id);
 }
